backend_n5/requirements.txt
flask
flask_mysqldb
flask-cors
PyJWT